requirements.txt
selenium
pytest
pytest-html


webdriver-manager
allure-python-commons